Fix gaps in tiles when they aren't aligned.
Non pixel-perfect tile alignment, where there are no lower level tiles, was showing slight gaps between the tiles.  The code to crop right and bottom edges was being invoked, because the texture coordinates of an unaligned tile occasionally are greater than 1.

To test this would require a new selenium test, since it is purely an artifact of how textures are rendered on quads in the vgl renderer.