Use firefox and xvfb to test webgl and examples.
To test this locally, make sure you have xvfb, firefox, and mesa installed.  The ci tests will no longer work under windows, as some environment variables are set in the npm commands.

When you run 'npm run start' and visit 127.0.0.1:9876/debug.html, all tests (both gl and non-gl) are run.  The ci tests run the unit tests in PhantomJS and the gl tests in Firefox.

There was a problem running the canvas tests in Chrome via karma debug, which has been fixed.

So far this has only two of the many selenium tests moved to headless testing.  It is enough to show the technique.