Improve some errors and some map coverage.
Throw errors, not strings.

Increase tests for the map class.

The map.fileReader function was documented in the geoJSON example as taking a fileReader instance, but it actually took a string and called createFileReader.  This function now can take either and has tests.

Change where coverage reports are written and collated from.  The web-gl and examples tests both use firefox, and their coverage clobbered each other.  We aren't using the examples tests to measure coverage, but rather than allow this to happen, change the coverage directories so that they no longer conflict.  Otherwise, if the tests were run in the order ff->examples->combine coverage, we would get erroneous coverage information.