Include svg layers in screenshots.
The screenshot function now returns a Promise.  It also can take an optional wait parameter to wait until the map is idle or just until a render animation frame has occurred.  An event is triggered when the screenshot is ready.

If we need it in the future, we can issue events during the preparation phase.

Update tests for promise-based screenshots.

A few of the selenium-based tests have been moved to the headless firefox-based tests.