Screenshot general html and svg layers.
This uses all css as part of the process.

Note that fonts look somewhat different in the screenshot -- specifically, svg foriegnObject is used to render the fonts, and this doesn't quite match the browser's native look.  Part of this is the lack of sub-pixel and hinting support (I think).

Screenshots of non-canvas layers will NOT work in IE.