Update the documentation in src/map.js.
Only the first 1/3 has been done, mainly to get feedback if this is what we want.

I've removed a few bits of vestigial code -- map.scale and map.maxZoom were unused.  The pan event claimed it included more information than it did, and included an extra empty element.

We could get rid of the map.origin() code -- we don't use it or test it other than simple coverage.

Perhaps we should have a more central place to document datatypes used in the documentation -- many are in the mapInteractor code.