More work on map documentation.
I've moved the typedefs in map and mapInteractor to their own file.

In principle, the documentation in src/map.js is done (barring typos).

I've fixed some documentation that was missing or wrong.  I rearranged some of the private inner functions to group them together and keep them above the initialization code.