Remove the map.resize method.
We marked the map.resize method as deprecated in version 0.6 (commit 27e6b36).  We completely ignore the x and y coordinates passed to it.  This change removes the resize and the internal m_x and m_y variables, moving the necessary functionality to the size method.

This also modifies the resize event by removing the x and y parameters.