Support annotation labels.
This adds a new feature type, `geo.textFeature`, where the data items
have a position, text, and styling.  Currently, this is only available
for the canvas renderer, and not all styles are hooked through.

Annotations now have `label`, `showLabel`, `labelStyle`, and
`description` properties.  The annotation layer has a `showLabels`
option for toggling all labels.

Update documentation of annotation.js and annotationLayer.js.

This needs to process and pass through styles to the renderer.  It needs
to generate a synthetic font style if the font substyles are used.  It
needs to support rotation, scaling, and offset on text features.  It
needs tests.  The annotation example should allow label adjustments,
including determining where placement is based from. Placement basis
needs to be an annotation option.