Add tutorial infrastructure.
This adds a method for creating tutorials.  There is a basic tutorial which is mostly defined in a pug template.  Code blocks are user-editable, and will update a live iframe with the results.  The pug template can define tests for code blocks which are run as part of the test suite.

Fixed some inconsistent use of jQuery $.Deferred where sometimes we added the `new` operator before calling it, which is unnecessary.