Add isObject utility function.  Use Array.isArray.
Fix jsdocs on all touched files.  Add tests.

Prior to this, we were using `instanceof Object` and `instanceof Array`.  For testing the upcoming tutorials branch, it is handy to be able to pass an object created in a parent window to a map running in an iframe.  However, `<parent object> instanceof Object` returned false in a map in the iframe, which makes it hard to interact with the map.  Similarly, the MDN docs recommend using `Array.isArray` over instanceof for this same reason.  Using instanceof to detect our own classes and Float32Array and similar structures still works without causing issues.  It is possible we will want to do something similar for detecting HTML Image and Canvas elements, but for now these have been left alone.