New layers' z-index could be wrong.
If a layer is deleted and then a new layer is added, the new layer's z-index can be the same as an existing layer.

This changes the default z-index logic to be one greater than any existing layer's z-index.  Also, if a layer is set explicitly to the same z-index of another layer, layers are shifted up one z-index value to ensure that all layers have unique z-indices.