Make a tutorial that acts as an editor for easy exploration.
This is my equivalent of using a fiddle site.  The code is stored as an ugly blob in the query string, so it is completely cookie and state free.

In general, any tutorial will store code in its url if 'keep=true' is added to the url when it is loaded.