Update examples.
Tweak a few examples to fix some inconsistencies and errors.  The goal is for every example to start in an interesting spot, and, moreover, for the thumbnail of the example to be representative of the example.

Added an optional `thumbquery` parameter to example.json files.  If present, this should be added to the base url when generating the thumbnail for the example.

Adjusted the zoom and center of several examples to start in more interesting locations.

Fixed some style issues.

If a heatmap was panned before it was ready, it could throw an error.

If a reprojection's viewport corners were ill-defined in the projection, the map would fail to draw.

For the WMS and SLD examples, switch the URL used for the geoserver to https.  This works -- the http server no longer does.

For the SLD example, add controls to allow setting the base url, layer name, and projection.  Clean up some of the pug code.  Update the thumbnail to match what the default settings show.