Add video quads.
Video quads are currently only supported by the canvas renderer.  If the video is not ready, a preview color can be shown, but not a preview image.

Added color quads to the canvas renderer.  This made things easier to test.

Improved some of the caching logic in quads to reduce reloading images and videos in some circumstances.

Added a utility method to invalidate one or all of the quad's caches.  This removes using a private method in the quad example.

Added code to defer rendering if something is busy.  This prevents flashing when seeking on a playing video.

Fixed an issue with the canvas transform on images and videos.