Make maps better at exiting and reloading.
Reduce repeated test code.  This reuses or replaces the #map node in the
headless tests rather than creating different named nodes for many
different tests.  This reduces the boiler-plate code for creating and
destroying maps in the tests.

Reusing elements exhibited some problems in map teardown and recreation.
Specifically: (a) When the map was exiting some functions could be
invoked with no-longer-valid renderer values.  (b) An animation frame
could fire after the vgl renderer had exited.  (c) The scene object
reset its children incorrectly.  (d) The sliderWidget turned off all
layer events for zoom rather than just its own.  (e) Tests that mock VGL
must clean up their maps before restoring VGL, or when the map is exited
it may not have the appropriate state information.

Fix some inconsistencies on widgets.  For instance, if a widget is
position relative to the screen and then changed to relative to the map,
it wouldn't update on pan, but if it started as relative to the map, it
would update on pan.

Update jsdocs on widgets.