Fix running headed chrome tests with proxy.
When we run headed tests on Chrome, we weren't invoking the proxy to
serve tiles locally and skip serving fonts.  This means that the tests
were running against internet data sources, and therefore subject to the
whims of connectivity and changes in OSM data.

Install Google fonts via an npm package and include them as local files
for the examples and tutorials.  This avoids having to fetch them each
time and also allows the headless tests to have them with certainty.

Fixed the webpack public path for examples and tutorials.