Update jquery.
This shouldn't result in any functional change.

The main difference between jQuery 2.x and jQuery 3.x, at least as
impacts GeoJS, is that `Deferred` objects are less synchronous.
Specifically, the `done`, `fail`, and `always` methods can be
synchronous if the deferred is resolved or rejected, but the `then`
method is never synchronous.  The major changes are in the tests which
relied on the synchronicity for running the tests; they no longer do so,
releasing time slices as appropriate.

Since GeoJS `object`s could take native promises, there is code to use
`always` if available and `then` if not.  In other places, we rely on
jQuery's `Deferred` having the `done` and `always` convenience
functions.