Improve resolving optional dependencies.
webpack makes any dependency that is not in a try-catch block a required
dependency.  Adjust code to be in conformance with this.  All require
statements for `d3` and `hammerjs` are now guarded.  Most `d3`
references are now via `d3Renderer.d3` to avoid having to have
conditionals in every file.