Speed up line style updates.
Before, any change to a line style would recompute the geometry
information as well.  Now, if the position of vertices has not changes,
the geometry is not recomputed.  Further, many style were being called
by function when they were their default values (such as the
strokeOffset not being set).  These properly are treated as fixed
values, avoiding many function calls.  The polygon feature, when
stroked, is better at passing constant values for stroke opacity to the
lines.

In practice, this results in a large speed up.  For a condition where
there are lines with ~100,000 vertices, the line _build call was reduced
from 450-500 ms to 12-15 ms.