Allow Hammer.js to exist as an npm-installed package
When resolving "require('hammerjs')", this will look for npm-installed
"hammerjs" packages to link to, in addition to a global "Hammer" variable.

Since "hammerjs" is an npm-optional dependency of GeoJS, if Webpack
rebundles GeoJS in a downstream project, it will be smart enough to elide
any "require('hammerjs')" statements in the built GeoJS. Without this
change, downstream Webpack is not able to elide "require('Hammer')" usages
(since the module name doesn't match the optional dependency name).