Fix testing with Chrome 65.
Chrome 65 was released 2018-3-6.  With it, some vertices appear odd in
GL rendered lines, but only on the headless GL or xvfb tests.  This
doesn't affect how the tests are rendered in Chrome 65 on a desktop, so
it is likely some sort of Chrome-65-with-mesa issue.  This is not a mesa
issue, since it was not updated.

Since this is just a testing artifact, rather than change the baselines,
relax the strictness of the image comparison.  Note that if travis
allowed pinning to a specific version of Chrome, I'd recommend pinning
to version 64 until Chrome fixes their issue.