Allow specifying the map's initial maxBounds in a different gcs.
If the bounds are known in a gcs other than the interface gcs, they had
to be converted first.

This change makes it easier to initialize a map with a stereographic
projection, where the unitsPerPixel can be specified to determine the
scope of the area shown (since it could be infinite).  For example,

var gcs = '+proj=stere +lat_0=90 +lat_ts=70 +lon_0=-45 +k=1 +x_0=0
+y_0=0 +a=6378273 +b=6356889.449 +units=m +no_defs'; var unitsPerPixel =
10000; geo.map({gcs: gcs, maxBounds: {left: -unitsPerPixel * 128, right:
unitsPerPixel * 128, bottom: -unitsPerPixel * 128, top: unitsPerPixel
* 128, gcs: gcs}, unitsPerPixel: unitsPerPixel });