Only allow pointer events on children of active layers.
This allows multiple active layers to work together.  Before, the
topmost active layer will receive all pointer events.

This also fixes an issue in the sceneObject exit function (it should
clear the parent, not delete the parent function).  This, in turn,
required a fix to the annotation test because the test map is destroyed
between tests.