Make the parameters passed to line style functions consistent.
When style or position functions are called for individual vertices on
polygons or lines on the canvas renderer, the functions are of the form
`styleFunc(vertex, vertexIndex, dataElement, dataIndex)`.  However, for
gl line features, this was of the form `styleFunc(vertex, vertexIndex,
lineFunc(dataElement, dataIndex), dataIndex)`.  The gl line feature has
been changed to make it consistent with other vertex style functions.