Allow changing the feature.mouseover order.
Add an event that allows changing the selection order for
feature.mouseover.  Also, add two common handlers for such ordering.

Specifically, the geo.event.feature.mouseover_order event is triggered
if the mouse is located on multiple elements of a feature.  The
`feature.mouseOverOrderHighestIndex` function will always sort these so
that the highest index element is on top.  The
`polygonFeature.mouseOverOrderClosestBorder` function will sort polygons
so that they are sorted based on how far the mouse is from their
borders, with the closest on top.  Note that, as always, the mouse must
be within the polygon for it to be considered.