Ensure that GL polygons are built once before their strokes.
When we build a webgl polygon we attach two vgl actors, one for the
polygon and one for the stroke (lines).  Whichever actor gets attached
first is drawn first.  We always want to draw the fill before the
stroke, but for efficiency it can be deferred an animation frame when
there are updates.  The first time this is rendered, don't allow such a
delay.  This change ensures that the fill actor is added before the
stroke actor.