Modify the webgl line feature shaders.
Chrome 65 and 66 when running in xvfb with openmesa do not compute
atan(0, <negative>) correctly.  Also, there are occasional int() casts
that round down even when they are small integers.  These changes make
image comparisons more robust by working around the atan issue (this may
be the issue detailed in
chromium.googlesource.com/angle/angle/commit/da9fb09).  I have submitted
a new issue to chromium:
https://bugs.chromium.org/p/chromium/issues/detail?id=841296 to track
this problem.

With this commit, the baselines were regenerated with Chrome 64, and the
test strictness requirements changed in PR #792 have been reverted.
There are subtly differences between Chrome 64 and 66 (font aliasing is
slightly different, and I think the other changes are because the shader
float precision is subtly different, though I'm not sure of that).