Add an isoline feature.
This adds a Math.log10 polyfill and removes the unused Math.sinh
polyfill.

Add simple isoline example.

Add an isoline tutorial.  This uses a data set of Oahu's elevation that
is lower resolution than the 'dense' set, primarily so that the tutorial
will run faster (and test faster).

This resolves #592.

The `used` function can be overridden for isoline and contour features.
This defaults to the `value` function returning a non-null, finite
number.

Pass the calculated position to the `value` function.  This allows the
value to be modified based on the position.

The isoline label feature is added to the list of dependent features so
that something like `myfeature.visible(false).draw()` does what is
expected.

Don't wrap across longitude if the data is more than 360 units wide,
since it probably isn't actually longitude in that case.

Reduce memory use for fully populated meshes.  For large meshes this can
reduce memory use and increase speed.  For small meshes, there is
practically no difference.

Label positions are calculated in map gcs, not interface gcs.  Lines are
not guaranteed to be linear in the interface gcs, so using it may put
labels off of the rendered line.

Modify layers for better nested support.

Layer's z-index should only be relative to their siblings, not all
layers.  Layers don't have to be children of the map; they can be
children of other layers.  Instead of enumerating the map layers for
determining siblings, enumerate the children of the layer's parent,
filtering to only those that are geo.layer instances.

This makes it easier to change the z index of a layer that has dependent
layers (such as the isoline feature's canvas layer for text).