Expose registries.
This, or some other way, is needed to that if a feature is available in
multiple renderers, so that a compound feature doesn't have to invoke a
secondary renderer if the current renderer is sufficient (see the change
for the annotationLayer, for instance).

This also has the virtue that the geo.registries entry can be inspected
to discover available features, layers, etc.