Add a parameter to the text feature to limit what gets rendered.
If specified, this skips rendering text that is outside of the current
viewport by a specified distance.  If you have very large text or text
with offsets, this could fail to render the text.  However, in
well-formed cases, this can skip rendering text that would not be
visible in any case.