Don't modify geojson object coordinates.
If a geojson object is passed to the annotation layer, the coordinates
used for polygon and line features referenced the arrays from the
geojson object.  These arrays could be modified.  For polygons, this
could remove a last point that duplicated a first point.  For both lines
and polygons, points are normalized to have x, y, z properties instead
of being stored in an array.  If the same geojson object was passed to
the annotation layer a second time, the polygon would always remove the
last point, even if it was not a duplicate of the first point because
the `[0]` and `[1]` properties of the points matched (as they no longer
exist, having been replaced with `x` and `y`).

This, instead, creates a new array for coordinates for polygons and
lines.  Although this uses more memory, it avoids the reuse problem.

Resolves #847.