Update documentation for pointFeature.
This also touches documentation on util/clustering.js, though that could
use some refactoring so that clusters are computed in map gcs rather
than in feature gcs (usually interface gcs) and the cluster size is
given in simpler units (such as display pixels).

This fixes a number of trivial documentation issues in other files, too.

Also, improve documentation of position in line and polygon features.