Fix the Deep Zoom example.
The image the Deep Zoom example used is offline.  Further, the example
hard-coded image sizes.  This uses an image from NASA, adds support for
reading and using the image metadata, and adds a test to ensure that the
example continues to work.

The example test infrastructure will wait for an idle function to be
available.