Update the fileReader and jsonReader docs.
Remove a function that was untested and incorrect from the fileReader.
Add tests for the fileReader.

Rename the `jsonReader` to `geojsonReader`.  An alias called
`jsonReader` is maintained to not break existing code.  The
geojsonReader can take style defaults for points, lines, and polygons.

This uses newer language features, such as the spread operator.

This resolves #822.  This resolves #297, though it does not add a load
event.  If that is still desired, a new issue should be made.