Use babel 6 and move where babel-polyfills are added.
Babel 7 is still pending official release and causes some issues.  Babel
7.0.0-rc.2 throws errors where 7.0.0-rc.1 does not.  Rather than figure
out what causes these, I recommend we wait until it is officially
released.  The main benefit of Babel 7 is that the polyfills can be
added conditionally based on actual use.

When adding babel-polyfill to src/index.js, the coverage for index.js is
wrong (the line numbers are incorrect).  Further, babel-polyfill is
fairly heavy, so just add it to the vendor.js (geo.js not geo.lean.js).
It also needs to be added to the test-utils to ensure that we can use
new language features in the tests with PhantomJS.  Reference to
babel-polyfill has been moved to src/polyfills.js, and that file has
been excluded from coverage analysis.

For the coverage reporters, disable the html reporter as it fails with
karma-coverage and babel.  One solution would be to switch to
karma-coverage-istanbul-reporter, but I don't think the html coverage is
strictly necessary as we use codecov.