Fix jsdoc of style.get and similar get functions.
jsdoc attempts to put `style.get` into the `geo.feature.style`
namespace, but `geo.feature.style` is a function, not a namespace, so it
doesn't appear.  This adds a jsdoc plugin that allows putting a dot in
the name of a doclet that appears in the rendered documentation.  This
results ins `style.get` appearing in the list of `geo.feature`'s
methods.