Separate static sections in the jsdocs.
Separate static and non-static methods and members in the jsdoc template
in navigation and container pages.  When there is a static (class)
method as well as an instance method, the navigation is confusing
otherwise.  Also, make sure the navigation is in the same order as the
container.  Before, typedefs were at the bottom of the container page,
but near the top of the navigation list.