Fix a possible polygon feature update after delete.
On a geo.gl.polygonFeature with stroke, the update can be delayed one
animation frame.  If the polygonFeature is deleted during that delay,
this could throw an error due to the styles of the feature being reset
as part of the deletion process.  This unschedules the update.

This PR also introduces a `ready` property to features to ensure deleted
features are not drawn.

This makes `m_this` references in the feature module more consistent.