Reduced use of the vgl module.
The `object` and `timestamp` classes were using the vgl module
definitions of these as base classes.  This obscured documentation --
the `modified` and `getMTime` methods were not in the geojs code, and
therefore did not appear in the jsdocs.  Rather than artificially inject
it, this simplifies code to move that functionality to the geojs repo.
This means that only the gl renderer has any dependency on the vgl
module (except fir the clustering code, which needs to be refactored in
any case).

As part of this, `getMTime` was renamed to the somewhat less obscure
`timestamp`.  The old name is still supported, but not in the jsdoc (it
wasn't in the jsdoc before, either).

This also converts all uses of `new Date().getTime()` to `Date.now()`.