Add an option to the tutorial editors to keep history in the browser.
If `keep=history` is in the tutorial's url, in addition to add the
source code to the window's location, the back button will return to the
previously source code of the tutorial.

I'm finding this is handy when using the editor to try out experimental
code, as those experiments are in the browser's history.  However, this
has been left as an option that has to be explicitly enabled, as if you
make many small edits, the history is very long and tedious without a
good indicator of when something significant or interesting occurred.