Test layer.js.
Move the layer specification to a typedef for better documentation.
Make sure all layer types use typedefs.

Fix issues with the layer id function.

Ensures that canvas and renderer can actually be specified when creating
a layer.  This allows layers to share a renderer or a canvas.  For the
webgl renderer, sharing a canvas results in flashing, as each layer
completely redraws the canvas.  When sharing a renderer, there is no
proper support for opacity (probably not for visibility) or z-order, but
this is a first step to implementing #942.

Removed some unused code.