Refactor how point clustering radius is used.
Before, it was cryptically linked to the display size, which was
counter-intuitive.  Now, the point clustering radius is based on display
pixels.

This also removes an unnecessary dependency on vgl from the clustering
algorithm.

Temporarily disable the test of the sld (WMS) example.  The external
server we use for that example is not responding.