Make zoom via wheel more responsive at the ends of the zoom range.
Prior to this, zooming via the wheel until the map was fully zoomed in
or zoomed out and then reversing the wheel could have a substantial lag
before the zoom would reverse.  While some debouncing is fine, this
sometimes felt unresponsive.  This ignores wheel events once the end of
the range is reached rather than adding more time to the transition.