Allow the fetch queue to have an initial size.
This adds the concept of an initial queue size to the fetch queue.  Once
the initial queue size requests have been satisfied, then the regular
queue size is used unless the initial queue size is reapplied.

For tile layers, the queue size and initial queue size can now be
specified directly, rather than reaching in to the _queue property and
modifying the queue there.

The default queue size remains at 6.  This value corresponds to most
browsers using a maximum of 6 requests per domain.  If a tile source
with subdomains is used, this could be increased to 6 * (number of
subdomains).  There doesn't seem to be any measurable improvement doing
so with the default osm tile source, though.

The default is not to use an initial queue size.  However, when using
tile servers that cache open tile sources and need time to serve the
first tile when not in cache, using an initial queue size of 1 can
reduce server load, as the tile server will get a single request, serve
it, and then get 6 simultaneous requests thereafter.