Improve tile overlap.
The common tile server with overlapping tiles is deep zoom.  Tiles in
the center of the image have overlap on all four sides, but tiles at the
edges don't overlap past the conceptual image space.  We properly
handled the right and bottom, but the top and left edges were expecting
tiles that were larger than deep zoom was providing.  For instance, a
256x256 tile with an overlap of 1 produces a 258x258 tile in the center,
but a 257x257 tile at the upper left and a 257x258 else on the left
edge.  This handles those properly.

The deepzoom example also maintains query parameters, making it easier
to modify for testing.