Improve webgl coordinate precision.
For line, point, polygon, and contour features, this uses a localized
origin for the view matrix to reduce jitter at high zoom levels.

Using global coordinates runs into the limitations of float precision,
which manes that small movements are functionally rounded by that
precision.  When panning at high zoom levels this manifests as jitter
where the features jump between discrete locations instead of panning
smoothly.  The default is to use the first coordinate of the data of the
feature for the origin.  For features that span a large area, this will
not help the jitter far from that coordinate (but won't be any worse
that before this PR).  The origin function is exposed on the style
object, so if a large area feature is needed, the application can bind
geo.event.pan to recompute the origin upon significant changes in the
map center at high zoom values.

Resolves #831.