Explicitly exit retired renderers.
Before, this was waiting for GC to clean up unused renderers, which
might not happen if renderers were created and destroyed without a major
GC event.  If GC didn't occur, the webgl context wouldn't be released.
This explicitly exits the renderers (and explicitly loses the webgl
context) as soon as the renderer is no longer in use (due to
autosharing).